Extract DockerClient abstraction
This will make easier for us to change the implmentation later.
